new prod

+++ nodes/VectorDBNode.py
@@ -1,6 +1,5 @@
 from elements.VideoElement import VideoElement
 from pymilvus import MilvusClient
-import json
 
 
 class VectorDBNode:
@@ -17,10 +16,10 @@ class VectorDBNode:
         # создание клиента и коллекции
         self.client = MilvusClient(uri=f"http://{self.host}:{self.port}")
         
-                # удаление бд по флагу
-        if self.drop_db:
-            res = self.client.delete(collection_name=self.collection_name, filter="id > 0")
-        
+        # удаление бд по флагу
+        if self.drop_db and self.client.has_collection(collection_name=self.collection_name):
+            _ = self.client.delete(collection_name=self.collection_name,  filter="id > 0")
+
         if not self.client.has_collection(collection_name=self.collection_name):
             self.client.create_collection(
                 collection_name=self.collection_name,
@@ -42,7 +41,7 @@ class VectorDBNode:
             res = res[0][0]
             video_element.top_1_neighbour_uuid = res["entity"]["uid"]
             metric_distance = float(res["distance"])
-            video_element.is_dublicate = metric_distance >= self.cosine_distance_treshold
+            video_element.is_dublicate = metric_distance <= self.cosine_distance_treshold
         return video_element
 
     def process_insert(self, video_element: VideoElement):
